requirements.txt
streamlit
pytesseract
pdf2image
Pillow
openpyxl
pandas
